Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 
 int main() {
-    float a=5;
-    float b=10;
-    float c=15;
+    int a=5;
+    int b=10;
+    int c=15;
     scanf("%d",&a);
     scanf("%d",&b);
     scanf("%d",&c);
-    printf("Average: %.2f", a+b+c/3);
+    printf("Average: %.2d", a+b+c/3);
     return 0;
 }
